requirements.txt
asyncpg>=0.25.0
SqlAlchemy[asyncio]>=1.4.45
SQLAlchemy-FullText-Search>=0.3.0
SQLAlchemy-Utils>=0.39.0
python-magic>=0.4.18
natsort>=7.1.1
aiohttp>=3.7
oauthlib>=3.2
packaging>=21.3
